+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.1.1 (2023-06-19)
+
+### Added
+
+- FilterMapCtx iterator to filter and map items simultaneously.
+- CtxMap adaptor to map a function over the context.
+
 ## v0.1.0 (2023-06-17)
 
 Initial proof of concept release.